feat: actualiza APIs acorde a cambios en servicio original

+++ Infrastructure/LineasRepository.cs
@@ -107,8 +107,8 @@ namespace SqlServerAdapter
                                 JOIN ssf.estadospais g ON e.id_estado=g.id_estado
                                 JOIN ssf.estadospais h ON f.id_estado=h.id_estado
                                 ,(
-                                	SELECT MAX(TRY_CAST(c.latitud AS FLOAT))+0.05 maxLatitud, MIN(TRY_CAST(c.latitud AS FLOAT))-0.05 minLatitud, 
-                                	       MAX(TRY_CAST(c.longitud AS FLOAT))-0.01 minLongitud, MIN(TRY_CAST(c.longitud AS FLOAT))+0.01 maxLongitud 
+                                	SELECT MAX(TRY_CAST(c.latitud AS decimal(10,5)))+0.05 maxLatitud, MIN(TRY_CAST(c.latitud AS decimal(10,5)))-0.05 minLatitud, 
+                                	       MAX(TRY_CAST(c.longitud AS decimal(10,5)))-0.01 minLongitud, MIN(TRY_CAST(c.longitud AS decimal(10,5)))+0.01 maxLongitud 
                                 	FROM ssf.rutas a
                                 	JOIN ssf.itinerario b ON a.id_ruta=b.id_ruta
                                 	JOIN ssf.puntospatrullaje c ON b.id_punto=c.id_punto
